+++ Vondera/Views/MarketPlaces/MarketPlaceReport.swift
@@ -47,9 +47,9 @@ struct MarketPlaceReport: View {
             
             if !isLoading {
                 Section("Sales & Profit") {
-                    ReportCard(title: "Total Orders", amount: items.getValidOrders().count, prefix: "Orders", desc : "Number of orders created in the time range")
+                    ReportCard(title: "Total Orders", amount: items.getValidOrders().count.double(), prefix: "Orders", desc : "Number of orders created in the time range")
                     
-                    ReportCard(title: "Products Sold", amount: items.getProductsCount(), prefix: "Products", desc : "Number of products sold")
+                    ReportCard(title: "Products Sold", amount: items.getProductsCount().double(), prefix: "Products", desc : "Number of products sold")
                     
                     ReportCard(title: "Sales", amount: items.totalSales(), prefix: "EGP", desc: "The Quantity of each product x The price of the product")
                     
@@ -57,14 +57,14 @@ struct MarketPlaceReport: View {
                     
                     ReportCard(title: "Orders Profit", amount: items.totalNetProfit(), prefix: "EGP", desc: "Profit calculated using this formula (Product Price + the shipping fees of the client) - (Discount on the order - Courier Fees - Seller Comission) \n * Make sure you add your courier with their fees")
                     
-                    ReportCard(title: "Delivery Success Rate", amount: items.getSuccessPercentage(), prefix: "%", desc: "(Delivered Order / Total Orders finished) * 100")
+                    ReportCard(title: "Delivery Success Rate", amount: items.getSuccessPercentage().double(), prefix: "%", desc: "(Delivered Order / Total Orders finished) * 100")
                 }
                 
                 
                 Section("On Going Orders") {
-                    ReportCard(title: "Pending Orders", amount: (items.getByStatue(statue: "Pending").count + items.getByStatue(statue: "Confirmed").count + items.getByStatue(statue: "Assembled").count), prefix: "Orders", desc : "Those are the orders that still with you")
+                    ReportCard(title: "Pending Orders", amount: (items.getByStatue(statue: "Pending").count + items.getByStatue(statue: "Confirmed").count + items.getByStatue(statue: "Assembled").count).double(), prefix: "Orders", desc : "Those are the orders that still with you")
                     
-                    ReportCard(title: "With Courier Orders", amount: items.getByStatue(statue: "Out For Delivery").count, prefix: "Orders", desc : "Those are the orders with your couriers")
+                    ReportCard(title: "With Courier Orders", amount: items.getByStatue(statue: "Out For Delivery").count.double(), prefix: "Orders", desc : "Those are the orders with your couriers")
                     
                     ReportCard(title: "Cash With Couriers", amount: items.getByStatue(statue: "Out For Delivery").totalCODAfterCourier(), prefix: "EGP", desc: "The amount of cash that expected to receive from the coureirs (Order price - Courier Fees)")
                 }
